requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
pydantic==2.12.5
spotipy==2.26.0
pandas==3.0.2
scikit-learn==1.8.0
python-dotenv==1.1.0
twilio==9.10.4
